Registration.cpp is complete except for connecting people, ie: struct Student has member 'Person caregiverPrimary'. This cannot be assigned as of now, me or Mohammed will add it in the future when most groundwork has been done. (Same with other structs with similar members)
StorageHandling has a global function called getUsernames(), which has an argument as 'vector<string>&' and can be used wherever the StorageHandling.h file is included, it will add all usernames from all saved filepaths to itself, will NOT remove any existing usernames stored in the input variable.

More similar functionality will be added as necessary.

Me and Mohammed went over it in a call, changing some of the Introduction.cpp file so it fit better in together with everything else (It was using the function name main(), which we already had in Source.cpp)